// shared/calculators/interfaces/config_interfaces.ts
export interface RectTransformationConfig {
  scaleX?: number; // Scaling factor along the side of a rotated rect that
  // was aligned with the X and Y axis before rotation
  // respectively.
  scaleY?: number;
  rotation?: number; // Additional rotation (counter-clockwise) around the
  // rect center either in radians or in degrees.
  rotationDegree?: number;
  shiftX?: number; // Shift along the side of a rotated rect that was
  // aligned with the X and Y axis before rotation
  // respectively. The shift is relative to the length of
  // corresponding side. For example, for a rect with size
  // (0.4, 0.6), with shiftX = 0.5 and shiftY = -0.5 the
  // rect is shifted along the two sides by 0.2 and -0.3
  // respectively.
  shiftY?: number;
  squareLong?: boolean; // Change the final transformed rect into a square
  // that shares the same center and rotation with
  // the rect, and with the side of the square equal
  // to either the long or short side of the rect
  // respectively.
  squareShort?: boolean;
}